requirements.txt
streamlit>=1.35
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3
prophet>=1.1.5
fpdf2>=2.7
joblib>=1.3
